+++ src/games/onset.rs
@@ -1,9 +1,10 @@
+use std::net::Ipv4Addr;
 use crate::{
     protocols::valve::{self, game, SteamApp},
     GDResult,
 };
 
-pub fn query(address: &str, port: Option<u16>) -> GDResult<game::Response> {
+pub fn query(address: &Ipv4Addr, port: Option<u16>) -> GDResult<game::Response> {
     let valve_response = valve::query(
         address,
         port.unwrap_or(7776),