+++ src/protocols/minecraft/protocol/legacy_v1_4.rs
@@ -1,3 +1,4 @@
+use std::net::Ipv4Addr;
 use crate::{
     bufferer::{Bufferer, Endianess},
     protocols::{
@@ -15,7 +16,7 @@ pub struct LegacyV1_4 {
 }
 
 impl LegacyV1_4 {
-    fn new(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<Self> {
+    fn new(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<Self> {
         let socket = TcpSocket::new(address, port)?;
         socket.apply_timeout(timeout_settings)?;
 
@@ -63,7 +64,7 @@ impl LegacyV1_4 {
         })
     }
 
-    pub fn query(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
+    pub fn query(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
         LegacyV1_4::new(address, port, timeout_settings)?.get_info()
     }
 }