+++ src/protocols/minecraft/protocol/mod.rs
@@ -1,3 +1,4 @@
+use std::net::Ipv4Addr;
 use crate::{
     protocols::minecraft::{
         protocol::{
@@ -24,7 +25,7 @@ mod legacy_v1_6;
 
 /// Queries a Minecraft server with all the protocol variants one by one (Java
 /// -> Bedrock -> Legacy (1.6 -> 1.4 -> Beta 1.8)).
-pub fn query(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
+pub fn query(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
     if let Ok(response) = query_java(address, port, timeout_settings.clone()) {
         return Ok(response);
     }
@@ -41,12 +42,12 @@ pub fn query(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>
 }
 
 /// Query a Java Server.
-pub fn query_java(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
+pub fn query_java(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
     Java::query(address, port, timeout_settings)
 }
 
 /// Query a (Java) Legacy Server (1.6 -> 1.4 -> Beta 1.8).
-pub fn query_legacy(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
+pub fn query_legacy(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<JavaResponse> {
     if let Ok(response) = query_legacy_specific(LegacyGroup::V1_6, address, port, timeout_settings.clone()) {
         return Ok(response);
     }
@@ -65,7 +66,7 @@ pub fn query_legacy(address: &str, port: u16, timeout_settings: Option<TimeoutSe
 /// Query a specific (Java) Legacy Server.
 pub fn query_legacy_specific(
     group: LegacyGroup,
-    address: &str,
+    address: &Ipv4Addr,
     port: u16,
     timeout_settings: Option<TimeoutSettings>,
 ) -> GDResult<JavaResponse> {
@@ -77,6 +78,6 @@ pub fn query_legacy_specific(
 }
 
 /// Query a Bedrock Server.
-pub fn query_bedrock(address: &str, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<BedrockResponse> {
+pub fn query_bedrock(address: &Ipv4Addr, port: u16, timeout_settings: Option<TimeoutSettings>) -> GDResult<BedrockResponse> {
     Bedrock::query(address, port, timeout_settings)
 }