+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Protocols:
 1. Added standard and serde derives to `GatheringSettings`.
 
 ### Breaking:
+- Every function that used `address: &str` has been changed to `address: &Ipv4Addr`.
+
 Services:
 - Valve Master Query:
 1. Removed Filter and SearchFilters lifetimes and changed `&'a str` to `String` and `&'a [&'a str]` to `Vec<String>`